// fashion_web_admin_ui/src/app/api/admin-api.service.generated.ts
export const ADMIN_API_BASE_URL = new InjectionToken<string>('ADMIN_API_BASE_URL');
// ...
export class CategoryDTO implements ICategoryDTO {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;

    constructor(data?: ICategoryDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): CategoryDTO {
        data = typeof data === 'object' ? data : {};
        let result = new CategoryDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        return data;
    }
}

export interface ICategoryDTO {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
}
// ...
export class PageIndex implements IPageIndex {
    _value?: number | undefined;

    constructor(data?: IPageIndex) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this._value = _data["_value"];
        }
    }

    static fromJS(data: any): PageIndex {
        data = typeof data === 'object' ? data : {};
        let result = new PageIndex();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["_value"] = this._value;
        return data;
    }
}

export interface IPageIndex {
    _value?: number | undefined;
}

export class PageSize implements IPageSize {
    _value?: number | undefined;

    constructor(data?: IPageSize) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this._value = _data["_value"];
        }
    }

    static fromJS(data: any): PageSize {
        data = typeof data === 'object' ? data : {};
        let result = new PageSize();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["_value"] = this._value;
        return data;
    }
}

export interface IPageSize {
    _value?: number | undefined;
}

export enum PaymentMethod {
    COD = "COD",
    CreditCard = "CreditCard",
    Momo = "Momo",
    PayPal = "PayPal",
}

export enum PaymentStatus {
    Pending = "Pending",
    Success = "Success",
    Failed = "Failed",
    Refunded = "Refunded",
}

export class ProductResponseDTO implements IProductResponseDTO {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    price?: number;
    thumbnailUrl?: string | undefined;
    categoryId?: string;

    constructor(data?: IProductResponseDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.price = _data["price"];
            this.thumbnailUrl = _data["thumbnailUrl"];
            this.categoryId = _data["categoryId"];
        }
    }

    static fromJS(data: any): ProductResponseDTO {
        data = typeof data === 'object' ? data : {};
        let result = new ProductResponseDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["price"] = this.price;
        data["thumbnailUrl"] = this.thumbnailUrl;
        data["categoryId"] = this.categoryId;
        return data;
    }
}

export interface IProductResponseDTO {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    price?: number;
    thumbnailUrl?: string | undefined;
    categoryId?: string;
}

export class ProductResponseDTOApiResponse implements IProductResponseDTOApiResponse {
    success?: boolean;
    message?: string | undefined;
    statusCode?: number | undefined;
    data?: ProductResponseDTO;

    constructor(data?: IProductResponseDTOApiResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.success = _data["success"];
            this.message = _data["message"];
            this.statusCode = _data["statusCode"];
            this.data = _data["data"] ? ProductResponseDTO.fromJS(_data["data"]) : undefined as any;
        }
    }

    static fromJS(data: any): ProductResponseDTOApiResponse {
        data = typeof data === 'object' ? data : {};
        let result = new ProductResponseDTOApiResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["success"] = this.success;
        data["message"] = this.message;
        data["statusCode"] = this.statusCode;
        data["data"] = this.data ? this.data.toJSON() : undefined as any;
        return data;
    }
}

export interface IProductResponseDTOApiResponse {
    success?: boolean;
    message?: string | undefined;
    statusCode?: number | undefined;
    data?: ProductResponseDTO;
}

export class ProductResponseDTOPageResponse implements IProductResponseDTOPageResponse {
    items?: ProductResponseDTO[] | undefined;
    total?: number;
    pageSize?: PageSize;
    pageIndex?: PageIndex;
    totalPages?: number;

    constructor(data?: IProductResponseDTOPageResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["items"])) {
                this.items = [] as any;
                for (let item of _data["items"])
                    this.items!.push(ProductResponseDTO.fromJS(item));
            }
            this.total = _data["total"];
            this.pageSize = _data["pageSize"] ? PageSize.fromJS(_data["pageSize"]) : undefined as any;
            this.pageIndex = _data["pageIndex"] ? PageIndex.fromJS(_data["pageIndex"]) : undefined as any;
            this.totalPages = _data["totalPages"];
        }
    }

    static fromJS(data: any): ProductResponseDTOPageResponse {
        data = typeof data === 'object' ? data : {};
        let result = new ProductResponseDTOPageResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.items)) {
            data["items"] = [];
            for (let item of this.items)
                data["items"].push(item ? item.toJSON() : undefined as any);
        }
        data["total"] = this.total;
        data["pageSize"] = this.pageSize ? this.pageSize.toJSON() : undefined as any;
        data["pageIndex"] = this.pageIndex ? this.pageIndex.toJSON() : undefined as any;
        data["totalPages"] = this.totalPages;
        return data;
    }
}

export interface IProductResponseDTOPageResponse {
    items?: ProductResponseDTO[] | undefined;
    total?: number;
    pageSize?: PageSize;
    pageIndex?: PageIndex;
    totalPages?: number;
}

export class ProductResponseDTOPageResponseApiResponse implements IProductResponseDTOPageResponseApiResponse {
    success?: boolean;
    message?: string | undefined;
    statusCode?: number | undefined;
    data?: ProductResponseDTOPageResponse;

    constructor(data?: IProductResponseDTOPageResponseApiResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.success = _data["success"];
            this.message = _data["message"];
            this.statusCode = _data["statusCode"];
            this.data = _data["data"] ? ProductResponseDTOPageResponse.fromJS(_data["data"]) : undefined as any;
        }
    }

    static fromJS(data: any): ProductResponseDTOPageResponseApiResponse {
        data = typeof data === 'object' ? data : {};
        let result = new ProductResponseDTOPageResponseApiResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["success"] = this.success;
        data["message"] = this.message;
        data["statusCode"] = this.statusCode;
        data["data"] = this.data ? this.data.toJSON() : undefined as any;
        return data;
    }
}

export interface IProductResponseDTOPageResponseApiResponse {
    success?: boolean;
    message?: string | undefined;
    statusCode?: number | undefined;
    data?: ProductResponseDTOPageResponse;
}
// ...
export interface FileParameter {
    data: any;
    fileName: string;
}